refactor(admin): AdminClasses sous 300 lignes — éclatement complet (#G1 ≤300)
AdminClasses.vue : 700 -> 265. Composable useAdminClasses (état + filtres +
filteredClasses/stats + load/cache/refresh + viewClasseDetails) ; sous-composants
ClassesStatsCards et ClassesFilters (v-model filière/niveau/statut) avec leur CSS.
La vue orchestre header + états + grille <ClassCard v-for> (déjà extraite). CSS
distribué verbatim.

Dette préexistante conservée (parité) : ArrowPathIcon n'était pas importé dans la
vue d'origine (icône non rendue) — comportement inchangé, documenté.

Tests : useAdminClasses + ClassesStatsCards + ClassesFilters. 18/18 verts
(avec ClassCard non-régression). vite build vert. Fichiers ≤300.
