+++ src/views/admin/AdminClasses.vue
@@ -11,94 +11,19 @@
       </div>
 
       <!-- Statistics -->
-      <div v-if="!loading && !error" class="stats-grid">
-        <div class="stat-card border-l-blue">
-          <div class="stat-header">
-            <i class="fa fa-home stat-icon"></i>
-            <span class="stat-label">Total Classes</span>
-          </div>
-          <p class="stat-value">{{ stats.total }}</p>
-        </div>
-
-        <div class="stat-card border-l-green">
-          <div class="stat-header">
-            <i class="fa fa-user stat-icon"></i>
-            <span class="stat-label">Total Étudiants</span>
-          </div>
-          <p class="stat-value">{{ stats.totalEtudiants }}</p>
-        </div>
-
-        <div class="stat-card border-l-orange">
-          <div class="stat-header">
-            <span class="stat-icon">☷</span>
-            <span class="stat-label">Total Matières</span>
-          </div>
-          <p class="stat-value">{{ stats.totalMatieres }}</p>
-        </div>
-
-        <div class="stat-card border-l-purple">
-          <div class="stat-header">
-            <span class="stat-icon">☀</span>
-            <span class="stat-label">Classes Actives</span>
-          </div>
-          <p class="stat-value">{{ stats.actives }}</p>
-        </div>
-      </div>
+      <ClassesStatsCards v-if="!loading && !error" :stats="stats" />
 
       <!-- Filters -->
-      <div v-if="!loading && !error" class="filters-card">
-        <div class="filters-grid">
-          <div class="filter-item">
-            <label class="filter-label">
-              <i class="fa fa-home filter-icon"></i>
-              Filière
-            </label>
-            <select v-model="filters.filiere_id" @change="applyFilters" class="filter-select">
-              <option value="">Toutes les filières</option>
-              <option v-for="filiere in filieres" :key="filiere.id" :value="filiere.id">
-                {{ filiere.name || filiere.nom }}
-              </option>
-            </select>
-          </div>
-
-          <div class="filter-item">
-            <label class="filter-label">
-              <i class="fa fa-bars filter-icon"></i>
-              Niveau
-            </label>
-            <select v-model="filters.niveau_id" @change="applyFilters" class="filter-select">
-              <option value="">Tous les niveaux</option>
-              <option v-for="niveau in niveaux" :key="niveau.id" :value="niveau.id">
-                {{ niveau.name || niveau.nom }}
-              </option>
-            </select>
-          </div>
-
-          <div class="filter-item">
-            <label class="filter-label">
-              <i class="fa fa-check filter-icon"></i>
-              Statut
-            </label>
-            <select v-model="filters.statut" @change="applyFilters" class="filter-select">
-              <option value="">Tous</option>
-              <option value="active">Actives</option>
-              <option value="inactive">Inactives</option>
-            </select>
-          </div>
-
-          <div class="filter-actions">
-            <button
-              v-if="filters.filiere_id || filters.niveau_id || filters.statut"
-              @click="resetFilters"
-              class="btn-reset"
-              title="Réinitialiser tous les filtres"
-            >
-              <ArrowPathIcon class="w-4 h-4" />
-              Réinitialiser
-            </button>
-          </div>
-        </div>
-      </div>
+      <ClassesFilters
+        v-if="!loading && !error"
+        v-model:filiere-id="filters.filiere_id"
+        v-model:niveau-id="filters.niveau_id"
+        v-model:statut="filters.statut"
+        :filieres="filieres"
+        :niveaux="niveaux"
+        @apply="applyFilters"
+        @reset="resetFilters"
+      />
 
       <!-- Loading state -->
       <ContentLoader v-if="loading" text="Chargement des classes..." />
@@ -148,223 +73,23 @@
 </template>
 
 <script setup>
-import { ref, computed, onMounted } from 'vue'
-import { useRouter } from 'vue-router'
+/**
+ * Gestion des classes (admin). Orchestrateur (#G1 ≤300) : la donnée et la logique
+ * vivent dans useAdminClasses ; l'UI est composée de ClassesStatsCards,
+ * ClassesFilters et de la grille de ClassCard.
+ */
 import DashboardLayout from '@/components/layout/DashboardLayout.vue'
 import ClassCard from '@/components/admin/ClassCard.vue'
+import ClassesStatsCards from '@/components/admin/ClassesStatsCards.vue'
+import ClassesFilters from '@/components/admin/ClassesFilters.vue'
 import ContentLoader from '@/components/common/ContentLoader.vue'
-import { klassciService } from '@/services/klassci'
-import { readCache, writeCache } from '@/services/cache'
+import { useAdminClasses } from '@/composables/useAdminClasses'
 
-const router = useRouter()
-
-const classes = ref([])
-const filieres = ref([])
-const niveaux = ref([])
-const matieres = ref([])
-const loading = ref(false)
-const error = ref(null)
-
-const filters = ref({
-  filiere_id: '',
-  niveau_id: '',
-  statut: ''
-})
-
-
-// Computed: Filtered classes
-const filteredClasses = computed(() => {
-  let result = classes.value
-
-  if (filters.value.filiere_id) {
-    result = result.filter(c => c.filiere?.id === parseInt(filters.value.filiere_id))
-  }
-
-  if (filters.value.niveau_id) {
-    result = result.filter(c => c.niveau?.id === parseInt(filters.value.niveau_id))
-  }
-
-  if (filters.value.statut) {
-    const isActive = filters.value.statut === 'active'
-    result = result.filter(c => c.is_active === isActive)
-  }
-
-  return result
-})
-
-// Computed: Statistics
-const stats = computed(() => {
-  const list = Array.isArray(classes.value) ? classes.value : []
-  return {
-    total: list.length,
-    totalEtudiants: list.reduce((sum, c) => sum + (c.places_occupees || 0), 0),
-    totalMatieres: list.reduce((sum, c) => sum + (c.nb_matieres || 0), 0),
-    actives: list.filter(c => c.is_active).length
-  }
-})
-
-// Load classes
-async function loadClasses() {
-  // Check cache
-  const cached = readCache('admin_classes')
-  if (cached) {
-    classes.value = cached.classes
-    filieres.value = cached.filieres
-    niveaux.value = cached.niveaux
-    matieres.value = cached.matieres
-    loading.value = false
-    refreshInBackground()
-    return
-  }
-
-  loading.value = true
-  error.value = null
-
-  try {
-    console.log('[ADMIN] Chargement de toutes les classes...')
-
-    // Load all data in parallel
-    const [classesData, matieresData, structureData] = await Promise.all([
-      klassciService.getClasses(),
-      klassciService.getMatieres(),
-      klassciService.getStructure()
-    ])
-
-    const rawClasses = classesData || []
-    matieres.value = matieresData || []
-    filieres.value = structureData?.filieres || []
-    niveaux.value = structureData?.niveaux || []
-
-    console.log('[ADMIN] Classes:', rawClasses.length, 'Matières:', matieres.value.length)
-
-    // Enrich each class with counters
-    const enrichedClasses = await Promise.all(
-      rawClasses.map(async (classe) => {
-        try {
-          const etudiants = await klassciService.getClasseEtudiants(classe.id)
-          const nbEtudiants = etudiants?.length || 0
-
-          // Count all matieres for this class (same logic as TeacherClasses)
-          const nbMatieres = matieres.value.length
-
-          const placesTotales = classe.effectif_max ||
-                                classe.capacite ||
-                                classe.places_totales ||
-                                classe.effectif ||
-                                (nbEtudiants > 0 ? Math.max(nbEtudiants, 30) : 30)
-
-          return {
-            ...classe,
-            places_occupees: nbEtudiants,
-            places_totales: placesTotales,
-            nb_matieres: nbMatieres
-          }
-        } catch (err) {
-          console.warn(`[WARN] Impossible d'enrichir classe ${classe.id}:`, err.message)
-          return {
-            ...classe,
-            places_occupees: 0,
-            places_totales: 30,
-            nb_matieres: matieres.value.length
-          }
-        }
-      })
-    )
-
-    classes.value = enrichedClasses
-
-    // Save to cache
-    writeCache('admin_classes', {
-      classes: classes.value,
-      filieres: filieres.value,
-      niveaux: niveaux.value,
-      matieres: matieres.value
-    })
-  } catch (err) {
-    console.error('[ERREUR] Chargement classes admin:', err)
-    error.value = 'Impossible de charger les classes. Veuillez réessayer.'
-  } finally {
-    loading.value = false
-  }
-}
-
-// Refresh in background
-async function refreshInBackground() {
-  try {
-    console.log('[BACKGROUND] Rafraîchissement classes admin...')
-    const [classesData, matieresData, structureData] = await Promise.all([
-      klassciService.getClasses(),
-      klassciService.getMatieres(),
-      klassciService.getStructure()
-    ])
-
-    const rawClasses = classesData || []
-    matieres.value = matieresData || []
-    filieres.value = structureData?.filieres || []
-    niveaux.value = structureData?.niveaux || []
-
-    const enrichedClasses = await Promise.all(
-      rawClasses.map(async (classe) => {
-        try {
-          const etudiants = await klassciService.getClasseEtudiants(classe.id)
-          const nbEtudiants = etudiants?.length || 0
-          const nbMatieres = matieres.value.length
-          const placesTotales = classe.effectif_max || classe.capacite || (nbEtudiants > 0 ? Math.max(nbEtudiants, 30) : 30)
-
-          return {
-            ...classe,
-            places_occupees: nbEtudiants,
-            places_totales: placesTotales,
-            nb_matieres: nbMatieres
-          }
-        } catch (err) {
-          return {
-            ...classe,
-            places_occupees: 0,
-            places_totales: 30,
-            nb_matieres: matieres.value.length
-          }
-        }
-      })
-    )
-
-    classes.value = enrichedClasses
-
-    writeCache('admin_classes', {
-      classes: classes.value,
-      filieres: filieres.value,
-      niveaux: niveaux.value,
-      matieres: matieres.value
-    })
-  } catch (error) {
-    console.warn('[BACKGROUND] Erreur rafraîchissement:', error)
-  }
-}
-
-// Apply filters
-function applyFilters() {
-  console.log('[FILTERS] Filtres appliqués:', filters.value)
-}
-
-// Reset filters
-function resetFilters() {
-  filters.value.filiere_id = ''
-  filters.value.niveau_id = ''
-  filters.value.statut = ''
-  console.log('[FILTERS] Filtres réinitialisés')
-}
-
-// View class details
-function viewClasseDetails(classe) {
-  router.push({
-    name: 'classe-details',
-    params: { id: classe.id }
-  })
-}
-
-onMounted(() => {
-  loadClasses()
-})
+const {
+  filieres, niveaux, loading, error, filters,
+  filteredClasses, stats,
+  loadClasses, applyFilters, resetFilters, viewClasseDetails,
+} = useAdminClasses()
 </script>
 
 <style scoped>
@@ -395,154 +120,6 @@ onMounted(() => {
   font-size: 1rem;
   color: var(--text-secondary);
   margin-top: 0.25rem;
-}
-
-/* Stat cards */
-.stats-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.stat-card {
-  background: var(--bg-primary);
-  border-radius: 1rem;
-  padding: 1.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  border-left: 4px solid transparent;
-  transition: all 0.2s;
-}
-
-.stat-card:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.border-l-blue {
-  border-left-color: #3b82f6;
-}
-
-.border-l-green {
-  border-left-color: #10b981;
-}
-
-.border-l-purple {
-  border-left-color: #8b5cf6;
-}
-
-.border-l-orange {
-  border-left-color: #f59e0b;
-}
-
-.stat-header {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  margin-bottom: 0.75rem;
-}
-
-.stat-icon {
-  font-size: 1.25rem;
-  line-height: 1;
-}
-
-.stat-label {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-}
-
-.stat-value {
-  font-size: 2rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin: 0;
-}
-
-/* Widget card (Filters) */
-.filters-card {
-  background: var(--bg-primary);
-  border-radius: 1rem;
-  padding: 1.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  margin-bottom: 2rem;
-}
-
-.filters-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 1.5rem;
-  align-items: end;
-}
-
-.filter-item {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.filter-label {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: var(--text-secondary);
-}
-
-.filter-icon {
-  font-size: 1rem;
-  line-height: 1;
-}
-
-.filter-select {
-  width: 100%;
-  padding: 0.625rem 0.75rem;
-  border: 1px solid var(--border-color);
-  border-radius: 0.5rem;
-  background: var(--input-bg);
-  color: var(--text-primary);
-  font-size: 0.875rem;
-  cursor: pointer;
-  transition: all 0.2s;
-}
-
-.filter-select:hover {
-  border-color: var(--primary-color);
-}
-
-.filter-select:focus {
-  outline: none;
-  border-color: var(--primary-color);
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
-}
-
-.filter-actions {
-  display: flex;
-  align-items: flex-end;
-}
-
-.btn-reset {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.625rem 1rem;
-  background: var(--bg-secondary);
-  color: var(--text-secondary);
-  border: 1px solid var(--border-color);
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-  width: 100%;
-  justify-content: center;
-}
-
-.btn-reset:hover {
-  background: var(--bg-hover);
-  color: var(--text-primary);
-  transform: translateY(-1px);
 }
 
 /* Class cards */
@@ -673,19 +250,7 @@ onMounted(() => {
     font-size: 1.5rem;
   }
 
-  .stat-value {
-    font-size: 1.75rem;
-  }
-
   .classes-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .stats-grid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  .filters-grid {
     grid-template-columns: 1fr;
   }
